==> notebook/00_loading_loom_data_filtering_Kenyon_cells_py37.ipynb <==
import loompy
import numpy as np
import pandas as pd
import os
import shutil

# Data can be downloaded from http://scope.aertslab.org
# Data is listed under Publically Available -> Drosophila -> Brain
# -> Aerts_Fly_AdultBrain_Filtered_57k.loom


# See http://linnarssonlab.org/loompy/apiwalkthrough/index.html
# for an introduction to working with loom files

# Because .loom data is written to disk as it's altered for e.g.
# exploratory analysis, a fresh copy of the original data is
# copied to re-run any analyses

og_data = "../data/00_original_data/ORIGINAL_Aerts_Fly_AdultBrain_Filtered_57k.loom"
wrk_data = "../data/01_working_data/WORKING_Aerts_Fly_AdultBrain_Filtered_57k.loom"
if os.path.isfile(og_data):
    shutil.copyfile(og_data, wrk_data)

# Make a connection to the working data
# **** Validation is set to False
# Presumably - the dataformat for loom has been altered in the time since this dataset was generated
# Validation is turned off to open, will be reshaped to a valid form in filtering process.
ds = loompy.connect(wrk_data, validate=False)

# Get sense of data
ds[0:10, 0:10] 

ds.shape

# The list of attributes associated with each Column/CELL
ds.ca.keys()

# The list of attributes associated with each row/GENE
ds.ra.keys()

# Exploring the Column/CELL attribute ClusterID
# which contains the code for the cell type of 
# each individual cell. See csv 
# ../data/02_filtered_kenyon_cells/cluster_id_cell_type_definition_at_resolution2.csv

# Clutser ID - Annotation
#  8 - G-KC
# 22 - a/b-KC
# 28 - a'/b'-KC

ds.ca.ClusterID

ds.ca.ClusterID.shape

# the loom file contains 56902 cells
# Each cell has a unique ClusterID / Cell type assigned to it
# 9 Seperate clusterings / cell-type callings were performed
# We will work with the initial set!

cell_type_id = []

for row in ds.ca.ClusterID:
    cell_type_id.append(row[0])
# Assign new column attribute 
ds.ca.cell_type_id = cell_type_id

# Create a True / False ( 0 / 1) column/CELL attribute
# too pull out only those cells which are one of the 
# three kenyon cell subtypes

is_kc = []

for cell_type_id in ds.ca.cell_type_id:
    if cell_type_id in [8, 22, 28]:
        is_kc.append(1)
    else:
        is_kc.append(0)
        
ds.ca.is_kc = is_kc

# Generating metadata df
# The attributes associated with each column/CELL is the metadata
# we will want to use to explore the data visually later
ca_keys = ds.ca.keys()
ddict = {}
for key in ca_keys:
    ddict[key] = list(ds.ca[key])
metadf = pd.DataFrame(ddict)
# Only some of these values are of interest to us
keep = ['CellID', 'Age', 'Gender', 'Genotype', 'Replicate',
        'nGene', 'nUMI', 'cell_type_id', 'is_kc']

metadf = metadf[keep]

# Drop all non-Kenyon cells
metadf = metadf[metadf.is_kc == 1]

metadf = metadf.set_index("CellID")

metadf.head()

# Almost 3000 total kenyon cells between the three types
# Consistent with number of KC listed in manuscript
metadf.shape

metadf.to_csv("../data/02_filtered_kenyon_cells/metadata_kenyon_cells.csv")

# Need to generate a new loom file containing only those cells which
# are a kenyon cell.
with loompy.new("../data/02_filtered_kenyon_cells/ORIGINAL_kenyon_cells.loom") as dsout:  # Create a new, empty, loom file
    cells = ds.ca.is_kc != 0
    for (ix, selection, view) in ds.scan(items=cells, axis=1):
        dsout.add_columns(view.layers, col_attrs=view.ca, row_attrs=view.ra)

# close out the original loom file, ds
ds.close()

# There is excess data stored in the column/CELL and row/GENE attributes
# which we either don't need or have stored as metadata
# These need to be removed so we can store our data as 
# a csv

ds2 = loompy.connect("../data/02_filtered_kenyon_cells/ORIGINAL_kenyon_cells.loom", validate=False)

ds2[0:10, 0:10]

# Remove every row/GENE attribute except for 'GENE'
ra_list = ds2.ra.keys()
ra_list.remove('Gene')

for ra in ra_list:
    del  ds2.ra[ra]

# Remove every column/CELL attribute except for 'CellID'
ca_list = ds2.ca.keys()
ca_list.remove('CellID')

for ca in ca_list:
    del ds2.ca[ca]

ds2

# Want to export data out now that it will be compatible with tabular formats
# help(ds2.export)
ds2.export("../data/02_filtered_kenyon_cells/CLEAN_kenyon_cells.tab")

ds2.close()

# Load in newly exported data
df2 = pd.read_csv("../data/02_filtered_kenyon_cells/CLEAN_kenyon_cells.tab", sep='\t')

# Export process generates some issues
# Data cleaning
df2 = df2.drop("CellID", axis=1)

# Symbol is the index column name
# used in a downstream analysis package
df2['symbol'] = df2['Unnamed: 0']
df2 = df2.set_index('symbol')

# Remove columns which are "unnamed" and empty
# or "unnamed" and contain gene names now stored
# in 'symbol'
df2 = df2.loc[:, ~df2.columns.str.contains('^Unnamed')]

# Drop row which is formed with empty row called "Gene"
df2 = df2.drop("Gene", axis=0)

df2.head()

df2.to_csv("../data/02_filtered_kenyon_cells/CLEAN_NOTRANSFORM_kenyon_cells.csv")

df2.applymap(lambda x : np.log2(x+1)).to_csv("../data/02_filtered_kenyon_cells/CLEAN_LOG2TRANSFORM_kenyon_cells.csv")
