==> notebook/01_ICIM_analysis_py27.ipynb <==
import sys
import pandas as pd
import numpy as np
import hdbscan

# Load up code to run ICIM
# Available from : https://github.com/felixhorns/FlyPN
sys.path.append("../code/ICIM")
import sct
reload(sct)
# Used because of nature of ICIM library
pd.options.mode.chained_assignment = None  # default='warn'

# The input for ICIM is log2 transformed data in practice
# See manuscript, github
df = pd.read_csv("../data/02_filtered_kenyon_cells/CLEAN_LOG2TRANSFORM_kenyon_cells.csv")
df = df.set_index("symbol")

# Load up metadata associated with these cells
metadata = pd.read_csv("../data/02_filtered_kenyon_cells/metadata_kenyon_cells.csv")
metadata = metadata.set_index("CellID")

metadata.head()

# ICIM takes a filtered dataset as one of its arguments
# Filtering requirements are taken from an up-to-date tutorial
# from the hemberg lab's tutorial on seurat, to mimic the presumed
# level of filtering used for the original seurat analsyis
# Source https://hemberg-lab.github.io/scRNA.seq.course/seurat-chapter.html#normalization

# filtered_df
f_df = df.copy(deep=True)

# seuset <- CreateSeuratObject(
#     raw.data = counts(deng),
#     min.cells = 3, 
#     min.genes = 200
# )

# Check to make sure each gene is present in at least 3 cells
def check_min_num_cells(row):
    num_pos = 0
    for gene_count in row:
        if gene_count > 0:
            num_pos += 1
            if num_pos >= 3:
                return True
    return False
   

min_num_cells = f_df.apply(check_min_num_cells, axis=1)
f_df = f_df[min_num_cells]

# Check to make sure each cell has at least 200 genes
def check_min_num_genes(col):
    num_pos = 0
    for gene_count in col:
        if gene_count > 0:
            num_pos += 1
            if num_pos >= 200:
                return True
    return False

min_num_genes = f_df.apply(check_min_num_genes, axis=0)
f_df = f_df.loc[:, min_num_genes]

f_df.shape

# Load the ICIM analysis object with associated data
# See https://github.com/felixhorns/FlyPN/blob/master/analysis/GH146_Fig2.ipynb 
# for full featured example

reload(sct)
from scipy.cluster import hierarchy
myICIM = sct.ICIM(f_df, df, TFs=[], CSMs=[], exclude=[], N=100,
                  correlation_cutoff=0.5,
                  min_hits=3,
                  exclude_max=2,
                  dropout_rate_low=0.3,
                  dropout_rate_high=1.0,
                  metric="correlation",
                  stop_condition="linkage_dist",
                  N_stop=50,
                  linkage_dist_stop=0.2)

# Run step command to iterate over first run of ICIM algo
# QC / Status Check
myICIM.step("0", verbose=True)

# Run full ICIM pipeline
myICIM.calc(verbose=True)

genes_KC_ICIM = myICIM.get_all_markers()
print "Genes found by ICIM", len(genes_KC_ICIM)

# with open("../data/03_ICIM_analysis/KC_genes_ICIM.txt", 'w') as out:
#     for x in genes_KC_ICIM:
#         out.write(x + "\n")

genes_KC_ICIM = []
with open("../data/03_ICIM_analysis/KC_genes_ICIM.txt", 'r') as infile:
    for line in infile:
        genes_KC_ICIM.append(line.strip())

# Subset data to genes desired
# Twice filtered df
f2_df = f_df.loc[genes_KC_ICIM]

# Calculate TSNE
reload(sct)
from sklearn.manifold import TSNE
myTSNE = sct.TSNE(f2_df, df, metadata)
myTSNE.calc_TSNE(perplexity=10, learning_rate=250, early_exaggeration=4.0, method="exact", random_state=1)

myTSNE.df_libs = myTSNE.df_libs.loc[myTSNE.df.columns,:]

# Using HDBSCAN to call clusters automatically, generate metadata for them
clusterer = hdbscan.HDBSCAN(min_cluster_size=30, min_samples=3).fit(myTSNE.X_tsne)
labels_HDBSCAN = clusterer.labels_

print "Clusters", max(labels_HDBSCAN)+1

zero_axis = myTSNE.X_tsne[:,0]
one_axis = myTSNE.X_tsne[:,1]
myTSNE.df_libs['x'] = zero_axis
myTSNE.df_libs['y'] = one_axis
myTSNE.df_libs['hdb_clust'] = labels_HDBSCAN
myTSNE.df_libs.index.name = "CellID"

def gen_kc_type(row):
    if int(row.cell_type_id) == 8:
        return 'G-KC'
    elif int(row.cell_type_id) == 22:
        return 'a/b-KC'
    elif int(row.cell_type_id) == 28:
        return "a'/b'-KC"
    
myTSNE.df_libs['subtype'] = myTSNE.df_libs.apply(gen_kc_type, axis=1)

myTSNE.df_libs.head()

myTSNE.df_libs.to_csv("../data/03_ICIM_analysis/KC_ICIM_TSNE_data.csv")

pp = pd.read_csv("../data/03_ICIM_analysis/KC_ICIM_TSNE_data.csv")

df_list = []
for var in ['Age', 'Gender', 'Genotype']:
    tdf = pp.groupby(['hdb_clust', var])['subtype'].count().to_frame(name='num_cells').reset_index().set_index('hdb_clust')
    tdf['variable'] = tdf[var]
    tdf['variable_cat'] = var
    tdf = tdf.drop(var, axis=1)
    df_list.append(tdf)

histo = pd.concat(df_list)
histo = histo[['variable_cat', 'variable', 'num_cells']]

total_cells = pp.groupby(['hdb_clust'])['subtype'].count().to_frame(name='total_number_cells_in_cluster')

histo = histo.join(total_cells)

histo['proportion'] = histo.num_cells / histo.total_number_cells_in_cluster

histo.head()

histo.to_csv("../data/03_ICIM_analysis/KC_bias_data.csv")
