==> notebook/03_pulling_ICIM_synaptic_genes_for_expression_analysis_py3.ipynb <==
import pandas as pd

# Exploring cluster informative genes expressed at the synaptic terminal
# Genes which explained differences in clusters were generated and stored
# at ../data/03_ICIM_analysis/KC_genes_ICIM.txt

gene_list = []
with open("../data/03_ICIM_analysis/KC_genes_ICIM.txt", 'r') as f:
    for line in f:
        gene_list.append(line.strip())

# Using the gene list pulled above, a Gene Ontology Analysis was run with 
# http://www.geneontology.org/page/go-enrichment-analysis

# A list of 18 genes expressed at synaptic terminals was produced
# These 18 genes will be used to visualize their expression levels
# in each of the clusters defined

# list of synaptic genes from GO analysis
# ../data/04_synaptic_gene_expression_profiling/pantherGeneList.txt

synap_gene_list = []
with open("../data/04_synaptic_gene_expression_profiling/pantherGeneList.txt", 'r') as f:
    for line in f:
        synap_gene_list.append(line.strip().split('\t')[1].strip())

synap_gene_list

df = pd.read_csv("../data/03_ICIM_analysis/KC_ICIM_TSNE_data.csv", index_col='CellID')

count = pd.read_csv("../data/02_filtered_kenyon_cells/CLEAN_LOG2TRANSFORM_kenyon_cells.csv", index_col='symbol')

count.head()

pull_count_table = count.loc[synap_gene_list]

pull_count_table.head()

# Create a function to pull out expression counts for each gene of interest
def pull_down_count(row, gene):
            cell = row.name
            return(pull_count_table.loc[gene, cell])

for gene in synap_gene_list:           
    df[gene] = df.apply(pull_down_count, gene=gene, axis=1)

df.head()

df.to_csv("../data/04_synaptic_gene_expression_profiling/CLEAN_LOG2TRANSFORM_WITHEXPRESSION_kenyon_cells.csv")
