==> notebook/06_TSNE_generation_from_synapse_organizing_genes.ipynb <==
import pandas as pd
import numpy as np
import sys
import hdbscan

# Load up code to run ICIM
# Available from : https://github.com/felixhorns/FlyPN
sys.path.append("../code/ICIM")
import sct
reload(sct)
# Used because of nature of ICIM library
pd.options.mode.chained_assignment = None  # default='warn'

# Set up to pull out the list of genes in the pruned expression table
df = pd.read_csv("../data/02_filtered_kenyon_cells/CLEAN_LOG2TRANSFORM_kenyon_cells.csv", index_col='symbol')

# ended up with a random row nan whose entire row is 0 counts - FAKE GENE
print('Shape of df with nan index : ' + str(df[df.index.isna()].shape))
print('total sum of expression values : ' + str(sum(df[df.index.isna()].iloc[0])))
print('drop row')
df = df[~df.index.isna()]

df.shape

# Will generate each list of genes seperately in case it is necc. to analyze them seperately.
# After creation will compile them all into a seperate list

gene_list = []

# Nicotinic receptor component subtype genes
# Keep the gene (index column) only for those genes which contain the 'nAChR'
# These genes are discussed at http://flybase.org/reports/FBgg0000202.html
nAChR = list(df[df.index.str.contains('nAChR', na=False)].index)

# Muscarinic receptor component subtype genes
# Keep the gene (index column) only for those genes which contain the 'mAChR'
# These genes are discussed at http://flybase.org/reports/FBgg0000187.html
mAChR = list(df[df.index.str.contains('mAChR', na=False)].index)

# Defective proboscis extension reponse genes
# Keep the gene (index column) only for those genes which contain the 'dpr'
# These genes are discussed at http://flybase.org/reports/FBgg0000529.html
dpr = list(df[df.index.str.contains('dpr', na=False)].index)

# Dpr interacting proteins
# These come in two batches - DIP & two unnamed genes ['CG31814', 'CG45781']
# These genes are discussed at http://flybase.org/reports/FBgg0000530
dip = list(df[df.index.str.contains('DIP-', na=False)].index) + list(df[df.index.isin(['CG31814', 'CG45781'])].index)

dscam = list(df[df.index.str.contains('Dscam', na=False)].index)

for geneset in [nAChR, mAChR, dpr, dip, dscam]:
    gene_list.extend(geneset)

# TSNE Projection will be completed with # genes:
len(gene_list)

# Generate a filtered df based on these genes
f_df = df.copy(deep=True)
f_df = f_df.loc[gene_list]

f_df.shape

# Check to make sure each gene is present in at least 3 cells
def check_min_num_cells(row):
    num_pos = 0
    for gene_count in row:
        if gene_count > 0:
            num_pos += 1
            if num_pos >= 3:
                return True
    return False
   

min_num_cells = f_df.apply(check_min_num_cells, axis=1)
f_df = f_df[min_num_cells]

f_df.shape

# This is a cell which has no expression for the associated synaptic organizing genes
# It has to be removed from both the unfiltered/filtered df AND the metadata df
f_df.loc[:,f_df.apply(sum) == 0].columns

# remove any cells which have zero counts for all genes in gene_list
df = df.drop('CTACACCTCAGAGGTG-DGRP-551_9d_r1', axis=1)
f_df = f_df.drop('CTACACCTCAGAGGTG-DGRP-551_9d_r1', axis=1)

df.shape

f_df.shape

# load up metadata that corresponds to full df
metadata = pd.read_csv("../data/02_filtered_kenyon_cells/metadata_kenyon_cells.csv")
metadata = metadata.set_index("CellID")

metadata = metadata.drop('CTACACCTCAGAGGTG-DGRP-551_9d_r1')

metadata.shape

# Calculate TSNE
reload(sct)
from sklearn.manifold import TSNE
myTSNE = sct.TSNE(f_df, df, metadata)
myTSNE.calc_TSNE(perplexity=10, learning_rate=250, early_exaggeration=4.0, method="exact", random_state=1)

# Using HDBSCAN to call clusters automatically, generate metadata for them
clusterer = hdbscan.HDBSCAN(min_cluster_size=30, min_samples=3).fit(myTSNE.X_tsne)
labels_HDBSCAN = clusterer.labels_

print "Clusters", max(labels_HDBSCAN)+1

zero_axis = myTSNE.X_tsne[:,0]
one_axis = myTSNE.X_tsne[:,1]
myTSNE.df_libs['x'] = zero_axis
myTSNE.df_libs['y'] = one_axis
myTSNE.df_libs['hdb_clust'] = labels_HDBSCAN


def gen_kc_type(row):
    if int(row.cell_type_id) == 8:
        return 'G-KC'
    elif int(row.cell_type_id) == 22:
        return 'a/b-KC'
    elif int(row.cell_type_id) == 28:
        return "a'/b'-KC"
    
myTSNE.df_libs['subtype'] = myTSNE.df_libs.apply(gen_kc_type, axis=1)
myTSNE.df_libs.to_csv("../data/05_synapse_organizing_gene_TSNE_gen/KC_synapse_organizing_gene_TSNE_data.csv")

myTSNE.df_libs.columns
