==> notebook/08_generating_datatables_for_shiny_py3.ipynb <==
import pandas as pd
import numpy as np

metadata = pd.read_csv("../data/03_ICIM_analysis/KC_ICIM_TSNE_data.csv")

metadata.head()

metadata['dna_index'] = metadata.CellID.str.split('-').str[0]

metadata = metadata.drop(['nGene', 'nUMI', 'cell_type_id', 'is_kc', 'x', 'y', 'CellID', 'Replicate'], axis=1)
metadata = metadata.set_index('dna_index')

metadata.head()

cell_number_info = pd.read_csv("../data/06_tss_data/indexed_barcodes.tsv", sep='\t', header=None)
cell_number_info.columns = ['cell_number', 'dna_index']
cell_number_info = cell_number_info.set_index("dna_index")

cell_number_info.columns = ['cell_number', 'dna_index']

metadata = metadata.join(cell_number_info)

metadata = metadata.set_index("cell_number")

metadata.head()

metadata.to_csv("../data/06_tss_data/umi_with_metadata.csv")

# Fix up UMI table by replacing flybase contigs with gene symbols
umi = pd.read_csv("../data/06_tss_data/all_umis_merged.tsv", sep="\t")
umi = umi.set_index('cell_number')

umi.head()

umi['facet'] = "Read Mappings"
umi = umi.drop(["strand", "numdup", "back_coord"], axis=1)

umi.head()

## Using this list of genes to convert to gene symbols
## done with this online tool, here 
## http://flybase.org/convert/id
## downloaded to ../data/06_tss_data/FlyBase_IDs.txt
with open("../data/06_tss_data/flybase_gene_id_list_to_convert.txt", 'w') as f:
    for fb in umi.contig.unique():
        f.write("{}\n".format(fb))

# generate map df
converter = pd.read_csv("../data/06_tss_data/FlyBase_IDs.txt", sep="\t")
converter = converter[['submitted_id', 'current_symbol']]
converter = converter.set_index('submitted_id')

# Make sure every gene is accounted for
umi[~umi.contig.isin(converter.index)]

# We're good!
converter.columns = ['symbol']

umi = umi.join(converter, on='contig')

umi = umi.drop("contig", axis=1)

umi.head()

umi.to_csv("../data/06_tss_data/all_umis_merged.csv")

## Fix up exon table as well
exon = pd.read_csv("../data/06_tss_data/exon_local_coords.tsv", sep='\t')
exon = exon.set_index("gene_id")
exon.head()

exon['coord'] = exon.local_lesser_coord
exon["back_coord"] = exon.local_greater_coord
exon = exon.drop(["exon", "local_lesser_coord", "local_greater_coord"], axis=1)
exon['facet'] = "Exon Map"
exon.head()

# Need a new gene list for exon genes
## done with this online tool, here 
## http://flybase.org/convert/id
## downloaded to ../data/06_tss_data/exon_FlyBase_IDs.txt
with open("../data/06_tss_data/flybase_exon_gene_id_list_to_convert.txt", 'w') as f:
    for fb in exon.index.unique():
        f.write("{}\n".format(fb))

# generate map df
exon_converter = pd.read_csv("../data/06_tss_data/exon_FlyBase_IDs.txt", sep="\t")
exon_converter = exon_converter[['submitted_id', 'current_symbol']]
exon_converter = exon_converter.set_index('submitted_id')

# Make sure every gene is accounted for
exon[~exon.index.isin(exon_converter.index)]

# We're good!
exon_converter.columns = ['symbol']

exon = exon.join(exon_converter)
exon.index.name ='contig'
exon = exon.set_index("symbol")
exon = exon[exon.index.isin(umi.symbol.unique())]

exon.to_csv("../data/06_tss_data/exon_local_coords.csv")
