==> notebook/11_tss_expression_matrix_ICIM.ipynb <==
import sys
import pandas as pd
import numpy as np
import hdbscan

# Load up code to run ICIM
# Available from : https://github.com/felixhorns/FlyPN
sys.path.append("../code/ICIM")
import sct
reload(sct)
# Used because of nature of ICIM library
pd.options.mode.chained_assignment = None  # default='warn'

%matplotlib inline
from matplotlib import pyplot as plt
import matplotlib as mpl

import seaborn as sns

# Original untransformed data, generated and saved here
# df = pd.read_csv("../data/08_tss_expression_matrix_ICIM/exon_expression_matrix.csv")
# df = df.set_index("Unnamed: 0")
# df.index.name = 'symbol'
#df['symbol'] = df['Unnamed: 0']
#df = df.set_index('symbol')
#df = df.drop('Unnamed: 0', axis=1)
#df.to_csv("../data/08_tss_expression_matrix_ICIM/log2_exon_expression_matrix.csv")

df.head()

df = pd.read_csv("../data/08_tss_expression_matrix_ICIM/log2_exon_expression_matrix.csv", index_col='symbol')

df.head()

# # Set up metadata
# cell_number_info = pd.read_csv("../data/06_tss_data/indexed_barcodes.tsv", sep='\t', header=None)
# cell_number_info.columns = ['cell_number', 'dna_index']
# cell_number_info = cell_number_info.set_index('dna_index')
# # Load up metadata associated with these cells
# metadata = pd.read_csv("../data/02_filtered_kenyon_cells/metadata_kenyon_cells.csv")
# metadata = metadata.set_index("CellID")
# metadata['dna_index'] = metadata.index.str.split('-').str[0]
# metadata = metadata.reset_index().set_index('dna_index')
# metadata = metadata.join(cell_number_info)
# metadata = metadata.reset_index().set_index("cell_number")
# # make csv
# metadata.to_csv("../data/08_tss_expression_matrix_ICIM/tss_metadata.csv")

df.index.unique()

df.shape

# ICIM takes a filtered dataset as one of its arguments
# Filtering requirements are taken from an up-to-date tutorial
# from the hemberg lab's tutorial on seurat, to mimic the presumed
# level of filtering used for the original seurat analsyis
# Source https://hemberg-lab.github.io/scRNA.seq.course/seurat-chapter.html#normalization

# filtered_df
f_df = df.copy(deep=True)

metadata = pd.read_csv("../data/08_tss_expression_matrix_ICIM/tss_metadata.csv", index_col="cell_number")
metadata = metadata[~metadata.index.duplicated(keep='first')]

# Pull out G-KC cells to cluster independently
gkc = list(metadata[metadata.cell_type_id == 8].index)
f_df = f_df.loc[:,f_df.columns.astype(int).isin(gkc)]

df = df.loc[:, df.columns.astype(int).isin(gkc)]

f_df.shape

df.shape

# Check to make sure each gene is present in at least 3 cells
def check_min_num_cells(row):
    num_pos = 0
    for gene_count in row:
        if gene_count > 0:
            num_pos += 1
            if num_pos >= 3:
                return True
    return False
   

min_num_cells = f_df.apply(check_min_num_cells, axis=1)
f_df = f_df[min_num_cells]

# Check to make sure each cell has at least 200 genes
def check_min_num_genes(col):
    num_pos = 0
    for gene_count in col:
        if gene_count > 0:
            num_pos += 1
            if num_pos >= 200:
                return True
    return False

min_num_genes = f_df.apply(check_min_num_genes, axis=0)
f_df = f_df.loc[:, min_num_genes]

f_df = f_df[f_df.apply(lambda x : sum(x) > 0, axis=1)]

f_df.shape

metadata = metadata.loc[f_df.columns.astype(int), :]
metadata.index.name = 'cell_number'

metadata.shape

metadata.head()

myDispersion = sct.dispersion(f_df)
myDispersion.calc_dispersion() # calculate overdispersion
genes_overdispersed = myDispersion.get_hits(N=500)
Y = f_df.loc[genes_overdispersed.index]

# Calculate PCA
from sklearn.decomposition import PCA
myPCA = sct.PCA(Y, df, 10) # number of significant PCs was determined using shuffling performed elsewhere in parallel fashion
myPCA.pca()
Y_pca = pd.DataFrame(myPCA.X_pca.T)

pca_df = pd.DataFrame(myPCA.X_pca)
pca_df['cell_number'] = myPCA.X.columns
pca_df = pca_df.set_index('cell_number')
pca_df.index = pca_df.index.astype(int)

myPCA.explained_variance_ratio_()

# FlyBase ID conversion
cv = pd.read_csv("../data/06_tss_data/exon_FlyBase_IDs.txt", sep="\t", index_col="submitted_id")
cv = cv[['current_symbol']]

cv.head()

top_genes = []
for gene_trans in myPCA.top_loaded_genes():
    gene = gene_trans.split('_')[0]
    top_genes.append(cv.loc[gene, 'current_symbol'])

with open("../data/06_tss_data/g-kc_top_pca_genes.txt", 'w') as f:
    for gene in top_genes:
        f.write('{}\n'.format(gene))

top_genes

col_titles = []
comp = 0
for expl in myPCA.explained_variance_ratio_():
    col_titles.append("Component_{}".format(comp))
    comp += 1

pca_df.columns = col_titles

clusterer = hdbscan.HDBSCAN(min_cluster_size=30, min_samples=3).fit(myPCA.X_pca)
labels_HDBSCAN = clusterer.labels_

pca_df = pca_df.join(metadata)
pca_df['hdb_clust'] = labels_HDBSCAN

pca_df.to_csv("../data/08_tss_expression_matrix_ICIM/G_KC_PCA.data.csv")

# Calculate TSNE
reload(sct)
from sklearn.manifold import TSNE
myTSNE_PCA = sct.TSNE(Y_pca, df, metadata)
myTSNE_PCA.calc_TSNE(perplexity=30, learning_rate=500)

# Using HDBSCAN to call clusters automatically, generate metadata for them
clusterer = hdbscan.HDBSCAN(min_cluster_size=30, min_samples=3).fit(myTSNE_PCA.X_tsne)
labels_HDBSCAN = clusterer.labels_

print "Clusters", max(labels_HDBSCAN)+1

myTSNE_PCA.df_libs.shape

myTSNE_PCA.X_tsne.shape

zero_axis = myTSNE_PCA.X_tsne[:,0]
one_axis = myTSNE_PCA.X_tsne[:,1]
myTSNE_PCA.df_libs['x'] = zero_axis
myTSNE_PCA.df_libs['y'] = one_axis
myTSNE_PCA.df_libs['hdb_clust'] = labels_HDBSCAN
myTSNE_PCA.df_libs.to_csv("../data/08_tss_expression_matrix_ICIM/G_KC_PCA_TSNE_data.csv")

# Load the ICIM analysis object with associated data
# See https://github.com/felixhorns/FlyPN/blob/master/analysis/GH146_Fig2.ipynb 
# for full featured example

reload(sct)
from scipy.cluster import hierarchy
myICIM = sct.ICIM(f_df, df, TFs=[], CSMs=[], exclude=[], N=100,
                  correlation_cutoff=0.5,
                  min_hits=3,
                  exclude_max=2,
                  dropout_rate_low=0.3,
                  dropout_rate_high=1.0,
                  metric="correlation",
                  stop_condition="linkage_dist",
                  N_stop=50,
                  linkage_dist_stop=0.2)

# Run step command to iterate over first run of ICIM algo
# QC / Status Check
myICIM.step("0", verbose=True)

# Run full ICIM pipeline
myICIM.calc(verbose=True)

genes_KC_ICIM = myICIM.get_all_markers()
print "Genes found by ICIM", len(genes_KC_ICIM)

df.shape
